kmk: pedantic warnings.
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@2001 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kmk/glob/glob.c
@@ -388,7 +388,7 @@ glob (pattern, flags, errfunc, pglob)
 	{
 	  /* Allocate working buffer large enough for our work.  Note that
 	    we have at least an opening and closing brace.  */
-	  int firstc;
+	  size_t firstc;
 	  char *alt_start;
 	  const char *p;
 	  const char *next;
